+++ rbx_reflection_database/CHANGELOG.md
@@ -1,7 +1,10 @@
 # rbx\_reflection_database Changelog
 
 ## Unreleased Changes
-* Updated to Roblox version 582.
+* Updated to Roblox version 583.
+* Added support for reflecting `QDir` properties. ([#322])
+
+[#322]: https://github.com/rojo-rbx/rbx-dom/pull/322
 
 ## 0.2.6+roblox-572
 * Updated to Roblox version 572.